Organize folders for easy importing
