+++ package.json
@@ -13,6 +13,7 @@
     "lucide-react": "^0.291.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
+    "react-hot-toast": "^2.4.1",
     "react-router-dom": "^6.18.0"
   },
   "devDependencies": {
