+++ src/components/DashboardWrapper.jsx
@@ -1,20 +1,20 @@
 import React from "react";
-import { Sidebar } from "./Sidebar";
+import { Sidebar } from "./sidebar/Sidebar";
 import { Header } from "./Header";
 import { Outlet } from "react-router";
 
 export function DashboardWrapper() {
   return (
     <div className="body">
-    <Header />
+      <Header />
       <main>
-      <Sidebar/>
+        <Sidebar />
         <section className="content">
           <div className="content-container">
             <Outlet />
           </div>
         </section>
       </main>
-   </div>
+    </div>
   );
 }
