+++ src/components/PrivateRoutes.jsx
@@ -1,28 +1,29 @@
 import React, { useEffect } from "react";
-import { Navigate, Outlet } from "react-router-dom";
+import { Navigate, Outlet, useNavigate } from "react-router-dom";
 
 export const PrivateRoutes = () => {
-  console.log("private Routes");
-  const currentUser = JSON.parse(localStorage.getItem("currentUser"));
+  const navigate = useNavigate();
 
-  if (currentUser) {
-    console.log("currentUser Validated");
-    const userData = JSON.parse(localStorage.getItem("userData"));
-    const user = userData.find(
-      (user) =>
-        user.email === currentUser.email &&
-        user.password === currentUser.password
-    );
+  useEffect(() => {
+    const currentUser = JSON.parse(localStorage.getItem("currentUser"));
 
+    if (currentUser) {
+      const userData = JSON.parse(localStorage.getItem("userData"));
+      const user = userData.find(
+        (user) =>
+          user.email === currentUser.email &&
+          user.password === currentUser.password
+      );
 
-    if (user.loginStatus) {
-        return <><Outlet/></>
-      }
-      else{
-        return <Navigate to='/login'/>
+      if (!user || !user.loginStatus) {
+        // Navigate to login if the user is not found or not logged in
+        navigate("/login");
       }
     }
-  return <Navigate to="/login" />;
+  }, []); // Empty dependency array ensures this effect runs only once
+
+
+  return <Outlet />;
 };
 
 export default PrivateRoutes;
