+++ src/components/sidebar/components/index.jsx
@@ -1,0 +1,3 @@
+export {AdminSidebar} from './AdminSidebar'
+export {CustomerSidebar} from './CustomerSidebar'
+export {CustomLink} from './CustomLink'
