+++ src/components/sidebar/utils/index.jsx
@@ -1,0 +1,2 @@
+export { handleLogout } from "./handleLogout";
+
