+++ src/pages/buy-load/BuyLoad.jsx
@@ -1,5 +1,5 @@
 import React, { useState } from "react";
-import { Sidebar } from "../../components/Sidebar";
+import { Sidebar } from "../../components/sidebar/Sidebar";
 import { Header } from "../../components/Header";
 
 export function BuyLoad() {
