+++ src/pages/currency/Currency.jsx
@@ -1,6 +1,6 @@
 import React, { useEffect, useState } from "react";
 import { CurrencyRow } from "./CurrencyRow";
-import { Sidebar } from "../../components/Sidebar";
+import { Sidebar } from "../../components/sidebar/Sidebar";
 import { Header } from "../../components/Header";
 
 export function Currency() {
