+++ src/pages/index.jsx
@@ -1,0 +1,9 @@
+export {Accounts} from './accounts/Accounts'
+export {BudgetApp} from './budget/BudgetApp'
+export {BuyLoad} from './buy-load/BuyLoad'
+export {CreateUser} from './createuser/CreateUser'
+export {Currency} from './currency/Currency'
+export {Overview} from './overview/Overview'
+export {SendMoney} from './send-money/SendMoney'
+export {TransactPage} from './transactPage/TransactPage'
+export {Goals}  from './Goals'
