+++ src/pages/send-money/SendMoney.jsx
@@ -1,4 +1,4 @@
-import { Sidebar } from "../../components/Sidebar";
+import { Sidebar } from "../../components/sidebar/Sidebar";
 import { Header } from "../../components/Header";
 
 export function SendMoney() {
